requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
geoalchemy2
geopandas
pandas
pydantic
requests
Pillow
torch
transformers
sentencepiece